+++ packages/web/src/app/api/(client)/client.ts
@@ -8,10 +8,6 @@ import {
     SearchResponse,
 } from "@/features/search";
 import {
-    FileSourceRequest,
-    FileSourceResponse,
-} from "@/features/search/types";
-import {
     FindRelatedSymbolsRequest,
     FindRelatedSymbolsResponse,
 } from "@/features/codeNav/types";
@@ -20,13 +16,16 @@ import {
     GetFilesResponse,
     GetTreeRequest,
     GetTreeResponse,
-} from "@/features/fileTree/types";
+    FileSourceRequest,
+    FileSourceResponse,
+} from "@/features/git";
 
 export const search = async (body: SearchRequest): Promise<SearchResponse | ServiceError> => {
     const result = await fetch("/api/search", {
         method: "POST",
         headers: {
             "Content-Type": "application/json",
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
         },
         body: JSON.stringify(body),
     }).then(response => response.json());
@@ -46,6 +45,9 @@ export const getFileSource = async (queryParams: FileSourceRequest): Promise<Fil
 
     const result = await fetch(url, {
         method: "GET",
+        headers: {
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
+        }
     }).then(response => response.json());
 
     return result as FileSourceResponse | ServiceError;
@@ -61,6 +63,7 @@ export const listRepos = async (queryParams: ListReposQueryParams): Promise<List
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
         },
     }).then(response => response.json());
 
@@ -72,6 +75,7 @@ export const getVersion = async (): Promise<GetVersionResponse> => {
         method: "GET",
         headers: {
             "Content-Type": "application/json",
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
         },
     }).then(response => response.json());
     return result as GetVersionResponse;
@@ -80,6 +84,9 @@ export const getVersion = async (): Promise<GetVersionResponse> => {
 export const findSearchBasedSymbolReferences = async (body: FindRelatedSymbolsRequest): Promise<FindRelatedSymbolsResponse | ServiceError> => {
     const result = await fetch("/api/find_references", {
         method: "POST",
+        headers: {
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
+        },
         body: JSON.stringify(body),
     }).then(response => response.json());
     return result as FindRelatedSymbolsResponse | ServiceError;
@@ -88,6 +95,9 @@ export const findSearchBasedSymbolReferences = async (body: FindRelatedSymbolsRe
 export const findSearchBasedSymbolDefinitions = async (body: FindRelatedSymbolsRequest): Promise<FindRelatedSymbolsResponse | ServiceError> => {
     const result = await fetch("/api/find_definitions", {
         method: "POST",
+        headers: {
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
+        },
         body: JSON.stringify(body),
     }).then(response => response.json());
     return result as FindRelatedSymbolsResponse | ServiceError;
@@ -96,6 +106,9 @@ export const findSearchBasedSymbolDefinitions = async (body: FindRelatedSymbolsR
 export const getTree = async (body: GetTreeRequest): Promise<GetTreeResponse | ServiceError> => {
     const result = await fetch("/api/tree", {
         method: "POST",
+        headers: {
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
+        },
         body: JSON.stringify(body),
     }).then(response => response.json());
     return result as GetTreeResponse | ServiceError;
@@ -104,6 +117,9 @@ export const getTree = async (body: GetTreeRequest): Promise<GetTreeResponse | S
 export const getFiles = async (body: GetFilesRequest): Promise<GetFilesResponse | ServiceError> => {
     const result = await fetch("/api/files", {
         method: "POST",
+        headers: {
+            "X-Sourcebot-Client-Source": "sourcebot-web-client",
+        },
         body: JSON.stringify(body),
     }).then(response => response.json());
     return result as GetFilesResponse | ServiceError;