+++ packages/web/src/app/api/(server)/tree/route.ts
@@ -1,12 +1,13 @@
 'use server';
 
-import { getTree } from "@/features/fileTree/api";
-import { getTreeRequestSchema } from "@/features/fileTree/types";
+import { getTree } from "@/features/git/getTreeApi";
+import { getTreeRequestSchema } from "@/features/git";
+import { apiHandler } from "@/lib/apiHandler";
 import { requestBodySchemaValidationError, serviceErrorResponse } from "@/lib/serviceError";
 import { isServiceError } from "@/lib/utils";
 import { NextRequest } from "next/server";
 
-export const POST = async (request: NextRequest) => {
+export const POST = apiHandler(async (request: NextRequest) => {
     const body = await request.json();
     const parsed = await getTreeRequestSchema.safeParseAsync(body);
     if (!parsed.success) {
@@ -19,5 +20,5 @@ export const POST = async (request: NextRequest) => {
     }
 
     return Response.json(response);
-}
+});